Add format extra.
Why? The way we use openapi we usually use it with bravado. Both use jsonschema,
but `bravado` uses `jsonschema[format]`. As a result, if `openapi` is installed
first, when `bravado` is installed, it will see that `jsonschema` is installed
and fail to install the `format` extras. For simplicity sake, we just want to
add that extra here to avoid issues with the order that dependencies are
required.